+++ src/components/Comment/components/CommentHeader.tsx
@@ -8,6 +8,8 @@ interface IProps {
   userCommunity: string | undefined;
   EllipsisButton: () => React.JSX.Element;
   CommentMetrics: () => React.JSX.Element;
+  isAdmin?: boolean;
+  isOp?: boolean;
 }
 
 function CommentHeader({
@@ -16,6 +18,8 @@ function CommentHeader({
   userName,
   EllipsisButton,
   CommentMetrics,
+  isAdmin,
+  isOp,
 }: IProps): React.JSX.Element {
   return (
     <XStack space="$2.5" alignItems="center" pb="$2">
@@ -23,6 +27,8 @@ function CommentHeader({
         userIcon={creatorAvatar}
         userName={userName}
         userCommunity={userCommunity}
+        isAdmin={isAdmin}
+        isOp={isOp}
       />
       <CommentMetrics />
       <View ml="auto" mr="$2.5">