+++ src/components/Common/Badge/AdminBadge.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import Badge from '@components/Common/Badge/Badge';
+
+export default function AdminBadge(): React.JSX.Element {
+  return <Badge color="$admin" text="Admin" />;
+}