+++ src/components/Common/Badge/OpBadge.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import Badge from '@components/Common/Badge/Badge';
+
+export default function OpBadge(): React.JSX.Element {
+  return <Badge color="$op" text="OP" />;
+}